♻️ refactor : 구간 서비스 클래스 분리
