+++ src/main/java/subway/application/LineService.java
@@ -13,12 +13,8 @@ import subway.domain.StationRepository;
 import subway.dto.LineCreateRequest;
 import subway.dto.LineEditRequest;
 import subway.dto.LineResponse;
-import subway.dto.RegisterSectionRequest;
 import subway.dto.SectionResponse;
 import subway.exception.LineNotFoundException;
-import subway.exception.SectionConstraintException;
-import subway.exception.SectionNotFoundException;
-import subway.exception.StationNotFoundException;
 
 @Transactional(readOnly = true)
 @Service
@@ -79,6 +75,13 @@ public class LineService {
         );
     }
 
+    private Station findStationById(final List<Station> stations, final Long stationId) {
+        return stations.stream()
+                .filter(station -> station.getId() == stationId)
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("노선 지하철 정보가 올바르지 않습니다."));
+    }
+
     @Transactional
     public void edit(final Long lineId, final LineEditRequest lineEditRequest) {
         Line line = findLineBy(lineId);
@@ -98,59 +101,5 @@ public class LineService {
     private Line findLineBy(final Long lineId) {
         return lineRepository.findById(lineId)
                 .orElseThrow(LineNotFoundException::new);
-    }
-
-    @Transactional
-    public void registerSection(final Long lineId, final RegisterSectionRequest registerSectionRequest) {
-        Line line = lineRepository.findByIdWithStation(lineId)
-                .orElseThrow(LineNotFoundException::new);
-
-        List<Station> stations = stationRepository.findAllById(
-                List.of(registerSectionRequest.getDownStationId(), registerSectionRequest.getUpStationId())
-        );
-
-        Station station = findStationById(stations, registerSectionRequest.getDownStationId());
-        Station upStation = findStationById(stations, registerSectionRequest.getUpStationId());
-
-        validateRegisterStation(line, station);
-
-        Section section = sectionRepository.findByStationAndLine(line.getDownStation(), line)
-                .orElseThrow(SectionNotFoundException::new)
-                .updateDownStation(station, upStation);
-
-        sectionRepository.save(section);
-    }
-
-    private void validateRegisterStation(final Line line, final Station downStation) {
-        sectionRepository.findByStation(downStation).flatMap(Section::getDownStation)
-                .flatMap(station -> sectionRepository.findByStationAndLine(station, line))
-                .ifPresent(station -> {
-                    throw new SectionConstraintException();
-                });
-    }
-
-    private Station findStationById(final List<Station> stations, final Long stationId) {
-        return stations.stream()
-                .filter(station -> station.getId() == stationId)
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("노선 지하철 정보가 올바르지 않습니다."));
-    }
-
-    @Transactional
-    public void deleteSection(final Long lineId, final Long stationId) {
-        Line line = lineRepository.findById(lineId).orElseThrow(LineNotFoundException::new);
-        Station station = stationRepository.findById(stationId).orElseThrow(StationNotFoundException::new);
-        List<Section> sections = sectionRepository.findAllByLine(line);
-
-        line.canDeleteSection(sections, station);
-
-        Section section = sections.stream()
-                .filter(s -> s.equalStation(station))
-                .findAny()
-                .orElseThrow(SectionNotFoundException::new);
-
-        line.updateDownStation(section.getUpStation().orElseThrow(SectionConstraintException::new));
-
-        sectionRepository.delete(section);
     }
 }
